parser: add sats and hdop support to CayenneLLP
The CayenneLLP GPS format does not support accuracy information, while
sensors commonly offer the number of satellites or hdop values. To
reflect those information additional data fields are required.

This commit adds support for both `sats` and `hdop` values by using the
CayenneLLP types AnalogIn and DigitalIn. To avoid interpreting values
meant for other purposes, the value reading is only enabled if AnalogIn6
is set to `3.03`, which is reasonably unlikely happen unintentionally.
If AnalogIn6 is set to `3.03` DigitalIn7 is interpret as `sats` and
`AnalogIn7` is interpret as `hdop`.

Signed-off-by: Paul Spooren <mail@aparcar.org>